refactor: make firebase db required
