+++ src/extensions/session-calendar/calendar-states.ts
@@ -1,6 +1,4 @@
 /** @module SessionCalendar */
-import { cert, getApps, initializeApp } from 'firebase-admin/app';
-import { Firestore, getFirestore } from 'firebase-admin/firestore';
 import { CalendarQueueExtension } from './calendar-queue-extension';
 import { cyan, yellow } from '../../utils/command-line-colors';
 import { BaseServerExtension } from '../extension-interface';
@@ -9,7 +7,8 @@ import { GuildId, GuildMemberId } from '../../utils/type-aliases';
 import LRU from 'lru-cache';
 import { environment } from '../../environment/environment-manager';
 import { restorePublicEmbedURL } from './shared-calendar-functions';
-import { Collection } from 'discord.js';
+import { Collection, Guild } from 'discord.js';
+import { firebaseDB } from '../../global-states';
 
 /**
  * @module Backups
@@ -29,11 +28,7 @@ class CalendarExtensionState {
     // full url to the public calendar embed
     publicCalendarEmbedUrl = restorePublicEmbedURL(this.calendarId);
 
-    constructor(
-        private readonly serverId: string,
-        private readonly serverName: string,
-        private readonly firebaseDB?: Firestore
-    ) {}
+    constructor(private readonly guild: Guild) {}
 
     /**
      * Returns a new CalendarExtensionState for the server with the given id and name
@@ -43,25 +38,17 @@ class CalendarExtensionState {
      * @param serverName
      * @returns CalendarExtensionState
      */
-    static async create(
-        serverId: string,
-        serverName: string
-    ): Promise<CalendarExtensionState> {
+    static async create(guild: Guild): Promise<CalendarExtensionState> {
         const firebaseCredentials = environment.firebaseCredentials;
         if (
             firebaseCredentials.clientEmail === '' &&
             firebaseCredentials.privateKey === '' &&
             firebaseCredentials.projectId === ''
         ) {
-            return new CalendarExtensionState(serverId, serverName);
+            return new CalendarExtensionState(guild);
         }
-        if (getApps().length === 0) {
-            initializeApp({
-                credential: cert(firebaseCredentials)
-            });
-        }
-        const instance = new CalendarExtensionState(serverId, serverName, getFirestore());
-        await instance.restoreFromBackup(serverId);
+        const instance = new CalendarExtensionState(guild);
+        await instance.restoreFromBackup(guild.id);
         return instance;
     }
 
@@ -110,10 +97,7 @@ class CalendarExtensionState {
      * @param serverId
      */
     async restoreFromBackup(serverId: string): Promise<void> {
-        if (this.firebaseDB === undefined) {
-            return;
-        }
-        const backupDoc = await this.firebaseDB
+        const backupDoc = await firebaseDB
             .collection('calendarBackups')
             .doc(serverId)
             .get();
@@ -138,9 +122,6 @@ class CalendarExtensionState {
      * Backs up the calendar config to firebase
      */
     private async backupToFirebase(): Promise<void> {
-        if (this.firebaseDB === undefined) {
-            return;
-        }
         const backupData: CalendarConfigBackup = {
             calendarId: this.calendarId,
             publicCalendarEmbedUrl: this.publicCalendarEmbedUrl,
@@ -150,9 +131,9 @@ class CalendarExtensionState {
                     .map(([key, LRUEntry]) => [key, LRUEntry.value])
             )
         };
-        this.firebaseDB
+        firebaseDB
             .collection('calendarBackups')
-            .doc(this.serverId)
+            .doc(this.guild.id)
             .set(backupData)
             .then(() =>
                 console.log(
@@ -161,7 +142,7 @@ class CalendarExtensionState {
                             timeZone: 'PST8PDT'
                         })
                     )} ` +
-                        `${yellow(this.serverName)}]\n` +
+                        `${yellow(this.guild.name)}]\n` +
                         ` - Calendar config backup successful`
                 )
             )
